Move -jar entry point into its own file
