+++ dd-java-agent/src/main/java/datadog/trace/bootstrap/AgentBootstrap.java
@@ -64,6 +64,10 @@ public final class AgentBootstrap {
     }
   }
 
+  public static void main(final String[] args) {
+    AgentJar.main(args);
+  }
+
   private static synchronized URL installBootstrapJar(final Instrumentation inst)
       throws IOException, URISyntaxException {
     URL ddJavaAgentJarURL = null;
@@ -184,39 +188,5 @@ public final class AgentBootstrap {
             + "' is located in '"
             + jarUrl
             + "'. Make sure you don't have this .class-file anywhere, besides dd-java-agent.jar");
-  }
-
-  /**
-   * Main entry point.
-   *
-   * @param args command line agruments
-   */
-  public static void main(final String... args) {
-    try {
-      System.out.println(getAgentVersion());
-    } catch (final Exception e) {
-      System.out.println("Failed to parse agent version");
-      e.printStackTrace();
-    }
-  }
-
-  /**
-   * Read version file out of the agent jar.
-   *
-   * @return Agent version
-   */
-  public static String getAgentVersion() throws IOException {
-    final StringBuilder sb = new StringBuilder();
-    try (final BufferedReader reader =
-        new BufferedReader(
-            new InputStreamReader(
-                thisClass.getResourceAsStream("/dd-java-agent.version"), StandardCharsets.UTF_8))) {
-
-      for (int c = reader.read(); c != -1; c = reader.read()) {
-        sb.append((char) c);
-      }
-    }
-
-    return sb.toString().trim();
   }
 }
