+++ dd-java-agent/src/main/java/datadog/trace/bootstrap/AgentJar.java
@@ -1,0 +1,35 @@
+package datadog.trace.bootstrap;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
+
+/** Entry point when running the agent as a sample application with -jar. */
+public final class AgentJar {
+  private static final Class<?> thisClass = AgentJar.class;
+
+  public static void main(final String[] args) {
+    try {
+      System.out.println(getAgentVersion());
+    } catch (final Exception e) {
+      System.out.println("Failed to parse agent version");
+      e.printStackTrace();
+    }
+  }
+
+  public static String getAgentVersion() throws IOException {
+    final StringBuilder sb = new StringBuilder();
+    try (final BufferedReader reader =
+        new BufferedReader(
+            new InputStreamReader(
+                thisClass.getResourceAsStream("/dd-java-agent.version"), StandardCharsets.UTF_8))) {
+
+      for (int c = reader.read(); c != -1; c = reader.read()) {
+        sb.append((char) c);
+      }
+    }
+
+    return sb.toString().trim();
+  }
+}
